@@ -17,48 +17,45 @@
         "ionic_webpack": "./webpack.config.js"
     },
     "dependencies": {
-        "@angular/common": "4.1.3",
-        "@angular/compiler": "4.1.3",
-        "@angular/compiler-cli": "4.1.3",
-        "@angular/core": "4.1.3",
-        "@angular/forms": "4.1.3",
-        "@angular/http": "4.1.3",
-        "@angular/platform-browser": "4.1.3",
-        "@angular/platform-browser-dynamic": "4.1.3",
-        "@ionic-native/camera": "3.12.1",
-        "@ionic-native/core": "3.12.1",
-        "@ionic-native/google-maps": "3.12.1",
+        "@angular/common": "4.4.4",
+        "@angular/compiler": "4.4.4",
+        "@angular/compiler-cli": "4.4.4",
+        "@angular/core": "4.4.4",
+        "@angular/forms": "4.4.4",
+        "@angular/http": "4.4.4",
+        "@angular/platform-browser": "4.4.4",
+        "@angular/platform-browser-dynamic": "4.4.4",
+        "@ionic-native/core": "4.3.0",
         "@ionic-native/in-app-browser": "^4.3.0",
         "@ionic-native/native-storage": "^4.3.0",
-        "@ionic-native/splash-screen": "3.12.1",
-        "@ionic-native/status-bar": "3.12.1",
-        "@ionic/storage": "2.0.1",
-        "@ngx-translate/core": "6.0.1",
-        "@ngx-translate/http-loader": "0.0.3",
-        "cordova-android": "^6.2.3",
-        "cordova-browser": "^4.1.0",
-        "cordova-ios": "^4.4.0",
-        "cordova-plugin-device": "^1.1.4",
+        "@ionic-native/splash-screen": "4.3.0",
+        "@ionic-native/status-bar": "4.3.0",
+        "@ngx-translate/core": "8.0.0",
+        "@ngx-translate/http-loader": "2.0.0",
+        "cordova-android": "^6.3.0",
+        "cordova-browser": "^5.0.0",
+        "cordova-ios": "^4.5.1",
+        "cordova-plugin-device": "^1.1.6",
         "cordova-plugin-inappbrowser": "^1.7.1",
         "cordova-plugin-nativestorage": "^2.2.2",
         "cordova-plugin-splashscreen": "^4.0.3",
-        "cordova-plugin-statusbar": "^2.2.2",
-        "cordova-plugin-whitelist": "^1.3.1",
-        "ionic-angular": "3.6.1",
+        "cordova-plugin-statusbar": "^2.2.3",
+        "cordova-plugin-whitelist": "^1.3.2",
+        "ionic-angular": "3.7.0",
         "ionic-plugin-keyboard": "^2.2.1",
         "ionicons": "3.0.0",
         "jwt-decode": "^2.2.0",
         "moment": "^2.18.1",
         "query-string": "^5.0.0",
-        "rxjs": "5.4.0",
+        "rxjs": "5.4.3",
         "sw-toolbox": "3.6.0",
-        "zone.js": "0.8.12"
+        "zone.js": "0.8.18"
     },
     "devDependencies": {
-        "@ionic/app-scripts": "2.1.4",
+        "@ionic/app-scripts": "3.0.0",
         "@types/jwt-decode": "^2.2.1",
         "@types/query-string": "^5.0.0",
-        "typescript": "2.3.4"
+        "typescript": "2.5.3"
     },
     "description": "An Ionic project",
     "cordova": {
